added dumptree function

+++ encode.c
@@ -1,11 +1,12 @@
 # include <stdint.h>
 # include <stdio.h>
+#include <unistd.h>
 # include "huffman.h"
 
 int main(void)
 {
 	FILE *fp = fopen("sample.txt", "r");
-	static uint64_t histogram[256];
+	static uint32_t histogram[256];
 	histogram[0] = 0x01;
 	histogram[255] = 0x01;
 	uint8_t temp;
